External lookups - Add handling for relationship fields
External lookups can now handle relationship marker fields.

First, the user will be required to select the target record and then, on selection, the standard 'Create relationship' popup will be opened with the selected target already in place.

Also included is a minor fix for the record dump for BnF lookups.