Rec editor - Fixes for date field controls and caching details about files
Fixed an issue where the help date links were being place beneath all the date inputs.

Fixed an issue with caching file details for the record editor to use in other function (like the Obfuscated ID and External URL)

Fixed an issue with the error handling related to calculating the extents of an image.