Add support for automatic translating via Deepl
A new 'Translate' button has been added to the translating popup for when translating fields (record free text + block text fields, and term labels + descriptions).
This button will take the selected language + the first input for the field and send them to Deepl for translating.
On success a new input field will be added with the languages prefix and the result added to it.