Resizing images for webpages
Adds a new feature to create a smaller jpeg version of large images to allow quicker loading for website/webpages that use them.
This feature requires the PHP module for Imagick to be installed on the current server.

To get this scaled down image the requesting URL needs to contain the parameter 'fullres=0', Heurist will then check if the image is over 500 KB and return either the scaled down image or the original image.
On an error, the original image will be returned.

The resized jpeg will also be scaled down to a maximum of 1000x1000 pixels.