Added Case conversion recode function
This function will convert the case for field values based on user selected details, this only works for freetext and blocktext field.
The basic options include:
 - Lowercase or capitalise all letters
 - Only capitalise the first letter and the letter following a fullstop, or
 - Only capitalise the first letter of each word (the default option)

A list of exceptions (words and phrases with punctuation) can be provided to override the selected operation, for example, to retain words that should be a particular way if the phrase/word should be camel cased.