Monthly server database stats
In the background, this will create a list of simple stats for all databases on the server (record counts, last record modification, etc...).
The stats will be stored within _DB_STATS, and sent to the Heurist main server (currently heuristref.net) and a copy stored there.
This function will be performed once a month.