Renamed server_scripts --> server_management, reorganised, replaced bash scripts with php
The aim is to greatly simplify and solidify server setup and management. Scripts no longer need to be moved to /srv/scripts. they can be run in situ. A bunch of bash scripts have been combined into a single php file (draft by ChatGPT to be verified), more work still to be done